+++ __init__.py
@@ -1,6 +1,19 @@
+import sys
+
+from . import py  # noqa: F401
 from .py import freeu_extreme, nodes, powernoise, sonar
 
+
+def blep_init():
+    bi = sys.modules.get("_blepping_integrations", {})
+    if "sonar" in bi:
+        return
+    bi["sonar"] = sys.modules[__name__]
+    sys.modules["_blepping_integrations"] = bi
+
+
 sonar.add_samplers()
+blep_init()
 
 NODE_CLASS_MAPPINGS = nodes.NODE_CLASS_MAPPINGS | {
     "SonarPowerNoise": powernoise.SonarPowerNoiseNode,