+++ py/utils.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import math
+from functools import partial
 
 import torch
 from comfy.model_management import device_supports_non_blocking
@@ -86,30 +87,150 @@ def tensor_to(
     return tensor.to(dest, non_blocking=non_blocking)
 
 
-def _quantile_norm_scaledown(noise: torch.Tensor, nq: torch.Tensor) -> torch.Tensor:
-    mv = noise.abs().max().detach().item()
-    return noise if mv == 0 else torch.where(noise.abs() > nq, noise * (nq / mv), noise)
+def _quantile_norm_scaledown(
+    noise: torch.Tensor,
+    nq: torch.Tensor,
+    *,
+    dim,
+    **_kwargs: dict,
+) -> torch.Tensor:
+    noiseabs = noise.abs()
+    mv = noiseabs.max(dim=dim, keepdim=True).clamp(min=1e-06)
+    return noise if mv == 0 else torch.where(noiseabs > nq, noise * (nq / mv), noise)
+
+
+def _quantile_norm_wave(
+    noise: torch.Tensor,
+    nq: torch.Tensor,
+    *,
+    preserve_sign: bool = False,
+    wave_function=torch.sin,
+    pi_factor: float = 0.5,
+    wrong_mode: bool = False,
+    **_kwargs: dict,
+) -> torch.Tensor:
+    if wrong_mode:
+        multiplier = 1.0 / ((math.pi * pi_factor) / nq)
+    else:
+        multiplier = 1.0 / (nq / (math.pi * pi_factor))
+    pos_mask = noise >= 0
+    neg_mask = ~pos_mask
+    result = torch.zeros_like(noise)
+    result[pos_mask] = wave_function(noise.mul(multiplier))[pos_mask]
+    result[neg_mask] = wave_function(noise.mul(multiplier))[neg_mask]
+    result *= nq
+    return result.copysign(noise) if preserve_sign else result
+
+
+def _quantile_norm_mode(
+    noise: torch.Tensor,
+    nq: torch.Tensor,
+    *,
+    dim: int | None,
+    decimals=1,
+    **_kwargs: dict,
+) -> torch.Tensor:
+    return torch.where(
+        noise.abs() > nq,
+        noise.round(decimals=decimals).mode(dim=dim, keepdim=True).values,
+        noise,
+    )
 
 
 quantile_handlers = {
-    "clamp": lambda noise, nq: noise.clamp(-nq, nq),
+    "clamp": lambda noise, nq, **_kwargs: noise.clamp(-nq, nq),
     "scale_down": _quantile_norm_scaledown,
-    "tanh": lambda noise, nq: noise.tanh().mul_(nq.abs()),
-    "tanh_outliers": lambda noise, nq: torch.where(
+    "tanh": lambda noise, nq, **_kwargs: noise.tanh().mul_(nq.abs()),
+    "tanh_outliers": lambda noise, nq, **_kwargs: torch.where(
         noise.abs() > nq,
         noise.tanh().mul_(nq.abs()),
         noise,
     ),
-    "sigmoid": lambda noise, nq: noise.sigmoid().mul_(nq.abs()).copysign(noise),
-    "sigmoid_outliers": lambda noise, nq: torch.where(
+    "sigmoid": lambda noise, nq, **_kwargs: noise.sigmoid()
+    .mul_(nq.abs())
+    .copysign(noise),
+    "sigmoid_outliers": lambda noise, nq, **_kwargs: torch.where(
         noise.abs() > nq,
         noise.sigmoid().mul_(nq.abs()).copysign(noise),
         noise,
     ),
-    "tenth": lambda noise, nq: torch.where(noise.abs() > nq, noise * 0.1, noise),
-    "half": lambda noise, nq: torch.where(noise.abs() > nq, noise * 0.5, noise),
-    "zero": lambda noise, nq: torch.where(noise.abs() > nq, 0, noise),
-    "reverse_zero": lambda noise, nq: torch.where(noise.abs() >= nq, noise, 0),
+    "sin": partial(_quantile_norm_wave, wave_function=torch.sin),
+    "sin_wholepi": partial(
+        _quantile_norm_wave,
+        wave_function=torch.sin,
+        pi_factor=1.0,
+    ),
+    "sin_keepsign": partial(
+        _quantile_norm_wave,
+        wave_function=torch.sin,
+        preserve_sign=True,
+    ),
+    "sin_wrong": partial(_quantile_norm_wave, wave_function=torch.sin, wrong_mode=True),
+    "sin_wrong_wholepi": partial(
+        _quantile_norm_wave,
+        wave_function=torch.sin,
+        pi_factor=1.0,
+        wrong_mode=True,
+    ),
+    "sin_wrong_keepsign": partial(
+        _quantile_norm_wave,
+        wave_function=torch.sin,
+        preserve_sign=True,
+        wrong_mode=True,
+    ),
+    "cos": partial(_quantile_norm_wave, wave_function=torch.cos),
+    "cos_wholepi": partial(
+        _quantile_norm_wave,
+        wave_function=torch.cos,
+        pi_factor=1.0,
+    ),
+    "cos_keepsign": partial(
+        _quantile_norm_wave,
+        wave_function=torch.cos,
+        preserve_sign=True,
+    ),
+    "cos_wrong": partial(_quantile_norm_wave, wave_function=torch.cos, wrong_mode=True),
+    "cos_wrong_wholepi": partial(
+        _quantile_norm_wave,
+        wave_function=torch.cos,
+        pi_factor=1.0,
+        wrong_mode=True,
+    ),
+    "cos_wrong_keepsign": partial(
+        _quantile_norm_wave,
+        wave_function=torch.cos,
+        preserve_sign=True,
+        wrong_mode=True,
+    ),
+    "atan": lambda noise, nq, **_kwargs: noise.atan().mul_(nq.abs() / (math.pi / 2)),
+    "tenth": lambda noise, nq, **_kwargs: torch.where(
+        noise.abs() > nq,
+        noise * 0.1,
+        noise,
+    ),
+    "half": lambda noise, nq, **_kwargs: torch.where(
+        noise.abs() > nq,
+        noise * 0.5,
+        noise,
+    ),
+    "zero": lambda noise, nq, **_kwargs: torch.where(noise.abs() > nq, 0, noise),
+    "reverse_zero": lambda noise, nq, **_kwargs: torch.where(
+        noise.abs() >= nq,
+        noise,
+        0,
+    ),
+    "mean": lambda noise, nq, *, dim, **_kwargs: torch.where(
+        noise.abs() > nq,
+        noise.mean(dim=dim, keepdim=True),
+        noise,
+    ),
+    "median": lambda noise, nq, *, dim, **_kwargs: torch.where(
+        noise.abs() > nq,
+        noise.median(dim=dim, keepdim=True).values,
+        noise,
+    ),
+    "mode_1dec": partial(_quantile_norm_mode, decimals=1),
+    "mode_2dec": partial(_quantile_norm_mode, decimals=2),
 }
 
 
@@ -124,41 +245,66 @@ def quantile_normalize(
     pow_fac: float = 0.5,
     strategy: str = "clamp",
     strategy_handler=None,
+    use_abs: bool = True,
 ) -> torch.Tensor:
+    if noise.numel() == 0:
+        return noise
+    if isinstance(quantile, (tuple, list)):
+        for q in quantile:
+            noise = quantile_normalize(
+                noise=noise,
+                quantile=q,
+                dim=dim,
+                flatten=flatten,
+                nq_fac=nq_fac,
+                pow_fac=pow_fac,
+                strategy=strategy,
+                strategy_handler=strategy_handler,
+                use_abs=use_abs,
+            )
+        return noise
     if quantile is None or quantile <= 0 or quantile >= 1:
         return noise
-    orig_shape = noise.shape
-    if isinstance(quantile, (tuple, list)):
-        quantile = torch.tensor(
-            quantile,
-            device=noise.device,
-            dtype=noise.dtype,
+    if not use_abs:
+        pos_mask = noise >= 0
+        neg_mask = ~pos_mask
+        result = torch.zeros_like(noise)
+        result[pos_mask] = quantile_normalize(
+            noise=noise[pos_mask],
+            quantile=quantile,
+            dim=dim,
+            flatten=flatten,
+            nq_fac=nq_fac,
+            pow_fac=pow_fac,
+            strategy=strategy,
+            strategy_handler=strategy_handler,
+            use_abs=True,
         )
-    qdim = dim
+        result[neg_mask] = quantile_normalize(
+            noise=noise[neg_mask],
+            quantile=quantile,
+            dim=dim,
+            flatten=flatten,
+            nq_fac=nq_fac,
+            pow_fac=pow_fac,
+            strategy=strategy,
+            strategy_handler=strategy_handler,
+            use_abs=True,
+        )
+        return result
+    orig_shape = noise.shape
     if noise.ndim > 1 and flatten:
-        if qdim is not None and qdim >= noise.ndim:
-            qdim = 1 if noise.ndim > 2 else None
-        if qdim is None:
-            flatdim = 0
-        elif qdim in {0, 1}:
-            flatdim = qdim + 1
-        elif 1 < qdim < 4:  # 2, 3
-            noise = noise.movedim(qdim, 1)
-            tempshape = noise.shape
-            flatdim = 2
-        else:
-            raise ValueError(
-                "Cannot handling quantile normalization flattening dims > 3",
-            )
+        flatnoise = noise.flatten(start_dim=dim)
     else:
-        flatdim = None
+        flatten = False
+        flatnoise = noise
     nq = torch.quantile(
-        (noise if flatdim is None else noise.flatten(start_dim=flatdim)).abs(),
+        flatnoise.abs(),
         quantile,
-        dim=-1,
+        dim=-1 if flatten else dim,
+        keepdim=True,
     )
-    nq_shape = tuple(nq.shape) + (1,) * (noise.ndim - nq.ndim)
-    nq = nq.mul_(nq_fac).reshape(*nq_shape)
+    nq = nq.mul_(nq_fac)
     handler = (
         quantile_handlers.get(strategy)
         if strategy_handler is None
@@ -166,13 +312,16 @@ def quantile_normalize(
     )
     if handler is None:
         raise ValueError("Unknown strategy")
-    noise = handler(noise, nq)
-    noise = noise.abs().pow(pow_fac).copysign(noise)
-    if flatdim is not None and qdim in {2, 3}:
-        return (
-            noise.reshape(tempshape).movedim(1, qdim).reshape(orig_shape).contiguous()
-        )
-    return noise
+    noise = handler(
+        flatnoise,
+        nq,
+        orig_noise=noise,
+        dim=dim,
+        flatten=flatten,
+    )
+    if pow_fac not in {0.0, 1.0}:
+        noise = noise.abs().pow_(pow_fac).copysign(noise)
+    return noise if noise.shape == orig_shape else noise.reshape(orig_shape)
 
 
 def normalize_to_scale(latent, target_min, target_max, *, dim=(-3, -2, -1)):